compliance-platform/requirements.txt
# Compliance Platform Dependencies
# FastAPI Backend

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0

# Caching & Task Queue
redis==5.0.1
celery==5.3.6

# HTTP Client (for integrations)
httpx==0.26.0
aiohttp==3.9.1

# AWS SDK
boto3==1.34.14

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Linting
black==24.1.0
ruff==0.1.13
mypy==1.8.0
